Move detect_env to common.
